Fix building the environment.
Previously, the environment was getting wiped and recreated every time
through the REPL.  This obviously meant you couldn't define new things
and then use them in a later statement.

This commit changes the ReaderT to a StateT and then populates it like
you'd expect.  When "define" is used, the new definition gets put into
the environment permanently.  When "let" is used, the bindings are only
local to its body.  And so forth.