Change let syntax.
The bindings portion of the let syntax was taking just a list of objects
where the evens were names and the odds were values.  Regular scheme
expects a list of pairs instead.  Use that syntax.

In other words we were accepting this:

(let (x 1 y 2) (+ 1 2))

And now we accept this:

(let ((x 1) (y 2)) (+ 1 2))