Add type predicate functions to the primitive environment.
These have to be done in haskell because we don't really have types in
scheme.  Maybe one day I'll think of some clever way around it, though.

The only one missing right now is null?, and that's because it can be
implemented in scheme once I have some other stuff written.