Remove pos? and neg? from Prim.hs, and sort of type equivalence.
pos? and neg? can be implemented in scheme, so there's no need to do it
in Haskell.  Then just have one type equivalence function named eqv? and
make it work like the scheme documentation sounds like it should.

And finally, change the error reporting in the prim environment to make
use of the new TypeMismatch exception.