Add tab completion for functions known when the repl starts.
This includes everything in the primitive environment and everything
from library.scm.  It does not include "write", but I want to do away
with it anyway because it's very special.

It does not include new things you define in the repl yourself.  I need
to figure out how to do that.