Let Main take programs on the command line.
All command line arguments are checked to see if they're a file that
exists.  If so, each is executed in its own environment and the results
printed out.  If any command line arguments are given, the REPL will not
be run.

This is mainly for purposes of testing, though it could be useful for
other things in the future.  Also, I might want to write tests in some
other way.