Big messy reorganization commit.
I wanted to put everything into a Scheme/ directory to make it easy to
write test cases.  And then while I was doing that I had to reorganize
the .cabal file to allow for that.  And then while I was doing that, I
added -Wall to the options so then I had to fix a lot of problems.

Also I removed Pretty.hs, but I wasn't using it anyway.