Add the variadic arg form of "define",
This is supposed to be supported for lambdas too, but for now I'm only
doing it for regularly defined stuff.

I also added the "list" function as a test case for this implementation,
and a lot of improved comments about evalBody.  I am starting to wonder
about the usefulness of that function in general.