Add the define-condition-type function to the base environment.
Along with adding the new Error and ErrorType LispVals, this is the
first step in adding exceptions.  Error is an actual instance of an
error condition, and ErrorType is the type of an error.

This new function creates other functions and adds them to the
environment.  The next step is adding a basic set of conditions, having
the Haskell code raise/handle them, and then adding support for raising
and handling them in scheme.

Hopefully more of this can be written in scheme.