Add support for another vararg syntax with define.
There's an additional syntax where a function can take several named
parameters, and then have another parameter after the dot to catch the
rest of the arguments.  I was only allowing for either named parameters
or the catch-all parameter, not both.

This significantly reworks how this branch of eval functions - I've
condensed the two into one branch instead of having to add a third.