Use parsec's built-in function for parsing numbers.
This strips the parser down to just about as small as I can get it,
which means it's about time to start adding stuff.  This additionally
gets me support for putting a + before a number (I don't know why anyone
would ever do that, but alright) and hex and octal constants.  The
latter two will not be in whatever weird form scheme like, but the more
common C-style.