Hook up a default exception handler.
This commit got a little out of control.  Oh well.

First, I got rid of execFile and execText.  Everything uses the eval
versions now, and those functions always return both a LispVal and an
environment.  This is because the LispVal could be an Error.

Then, I changed safeExec to catchHaskellExceptions and made it always
return an Error too.  In the event of an exception originating in
Haskell, that's an internal-error which we can do nothing with except
exit.

Finally, I added defaultExceptionHandler which figures out what kind of
exception has been received and takes appropriate action.  If there's no
exception, it doesn't do anything.  I then changed everywhere that a
top-level exists (in the REPL, in Main.hs, and in the tests) to call
this function.