Use the error message functions in errorConstrFn, when possible.
This means that in scheme when someone uses make-*-error, the error
messages will get the same formatting as they would if the error came
from Haskell.

Note that not all errors can use these functions, because some take
different numbers of arguments.  This is a problem as long as scheme
errors are only allowed to pass in one argument.