Add error condition predicate functions.
These are for figuring out if a condition object is of a certain type.
They should now work for the built in conditions as well as user-defined
conditions.  They do not yet work for subtyping (that is, if B is a
subtype of A, A? will return false for a B object).