Get tab completion for things defined in the REPL working.
Basically, the settings have to be updated every time through for the
new bindings.  haskeline doesn't provide any way to do that, so the
existing repl function needs to get blown up.  It's now just named "rep"
because it doesn't loop.  The looping happens outside it so runInputT
can be run with new settings.