Do a better job with error handling in errorConstrFn.
The change that motivated this was making it so in scheme, you could
create a type error with two arguments just like you could in Haskell.
That required changing the pattern matching, at which point I decided to
just break it all up into error-specific chunks so I can later change
the number of arguments for other errors.