Rework program state.
It's not enough to just have one environment with all bindings, and
that's the sole program state.  In order to support types (just
Conditions for the moment, maybe records later) we need one environment
for types and one environment for bindings.

With that, program state is more than just one thing and needs to be
made into a data type.  That then requires changes throughout to get a
particular element out of the state, etc.