Split error conditions into three types: CondTy, Condition, and Raised.
CondTy is a type and lives in its own environment.  A new CondTy is
created with define-condition-type.

Condition is an instance of a CondTy and is created by any of the
make-*-error functions.  By itself, a Condition doesn't do anything.
It's just a value like any other.

A Raised is a Condition that has been converted into an exception with
the raise function.  When a Raised is encountered in evaluation,
evaluation immediately stops and the Raised value is returned as the
result.  This passes up the call stack until a guard expression or the
top level is encountered.

The latter is responsible for most of the change in this patch.