Rewrite list-ref and list-tail to do error checking.
In particular, check that they weren't passed a negative number as an
index (or offset), that they weren't passed an index larger than the
list itself, and that an empty list isn't encountered when it shouldn't
be.

I also changed list-error to take two arguments - the type of thing that
happened (empty list, index out of range, etc.) and the function it
happened in.  I don't want to go too crazy with the types of errors.